Extract full match simulation into FullMatchSimulationService
Move simulateMatch, simulateMatches, getBenchPlayers, getLineupPlayers,
calculateMvp, and forfeit logic from MatchdayOrchestrator into a dedicated
FullMatchSimulationService. The orchestrator now delegates to either
AIMatchResolver (fast AI-vs-AI path) or FullMatchSimulationService (full
simulation for player-involved batches) — cleaner separation of concerns.
